+++ modules/nodegroup/launch_template.tf
@@ -1,211 +1,32 @@
-resource "aws_launch_template" "this" {
-  count = var.is_create_launch_template ? 1 : 0
-  name_prefix = local.prefix
-  description = "aws_launch_template for custom nodegroup"
-
-  ebs_optimized = var.ebs_optimized
-  image_id      = var.ami_id
-
-  # # Set on node group instead
-  # instance_type = var.launch_template_instance_type
-  key_name  = var.key_name
-  user_data = local.platform[var.platform].user_data
-
-  vpc_security_group_ids = var.vpc_security_group_ids
-
-  default_version         = var.launch_template_default_version
-  update_default_version  = var.update_launch_template_default_version
-  disable_api_termination = var.disable_api_termination
-  # Set on EKS managed node group, will fail if set here
-  # https://docs.aws.amazon.com/eks/latest/userguide/launch-templates.html#launch-template-basics
-  # instance_initiated_shutdown_behavior = var.instance_initiated_shutdown_behavior
-  kernel_id   = var.kernel_id
-  ram_disk_id = var.ram_disk_id
-
-  dynamic "block_device_mappings" {
-    for_each = var.block_device_mappings
-    content {
-      device_name  = block_device_mappings.value.device_name
-      no_device    = lookup(block_device_mappings.value, "no_device", null)
-      virtual_name = lookup(block_device_mappings.value, "virtual_name", null)
-
-      dynamic "ebs" {
-        for_each = flatten([lookup(block_device_mappings.value, "ebs", [])])
-        content {
-          delete_on_termination = lookup(ebs.value, "delete_on_termination", null)
-          encrypted             = lookup(ebs.value, "encrypted", null)
-          kms_key_id            = lookup(ebs.value, "kms_key_id", null)
-          iops                  = lookup(ebs.value, "iops", null)
-          throughput            = lookup(ebs.value, "throughput", null)
-          snapshot_id           = lookup(ebs.value, "snapshot_id", null)
-          volume_size           = lookup(ebs.value, "volume_size", null)
-          volume_type           = lookup(ebs.value, "volume_type", null)
-        }
-      }
-    }
-  }
-
-  dynamic "capacity_reservation_specification" {
-    for_each = var.capacity_reservation_specification != null ? [var.capacity_reservation_specification] : []
-    content {
-      capacity_reservation_preference = lookup(capacity_reservation_specification.value, "capacity_reservation_preference", null)
-
-      dynamic "capacity_reservation_target" {
-        for_each = try([capacity_reservation_specification.value.capacity_reservation_target], [])
-        content {
-          capacity_reservation_id = lookup(capacity_reservation_target.value, "capacity_reservation_id", null)
-        }
-      }
-    }
-  }
-
-  dynamic "cpu_options" {
-    for_each = var.cpu_options != null ? [var.cpu_options] : []
-    content {
-      core_count       = cpu_options.value.core_count
-      threads_per_core = cpu_options.value.threads_per_core
-    }
-  }
-
-  dynamic "credit_specification" {
-    for_each = var.credit_specification != null ? [var.credit_specification] : []
-    content {
-      cpu_credits = credit_specification.value.cpu_credits
-    }
-  }
-
-  dynamic "elastic_gpu_specifications" {
-    for_each = var.elastic_gpu_specifications != null ? [var.elastic_gpu_specifications] : []
-    content {
-      type = elastic_gpu_specifications.value.type
-    }
-  }
-
-  dynamic "elastic_inference_accelerator" {
-    for_each = var.elastic_inference_accelerator != null ? [var.elastic_inference_accelerator] : []
-    content {
-      type = elastic_inference_accelerator.value.type
-    }
-  }
-
-  dynamic "enclave_options" {
-    for_each = var.enclave_options != null ? [var.enclave_options] : []
-    content {
-      enabled = enclave_options.value.enabled
-    }
-  }
-
-  # Set on EKS managed node group, will fail if set here
-  # https://docs.aws.amazon.com/eks/latest/userguide/launch-templates.html#launch-template-basics
-  # dynamic "hibernation_options" {
-  #   for_each = var.hibernation_options != null ? [var.hibernation_options] : []
-  #   content {
-  #     configured = hibernation_options.value.configured
-  #   }
-  # }
-
-  # Set on EKS managed node group, will fail if set here
-  # https://docs.aws.amazon.com/eks/latest/userguide/launch-templates.html#launch-template-basics
-  # dynamic "iam_instance_profile" {
-  #   for_each = [var.iam_instance_profile]
-  #   content {
-  #     name = lookup(var.iam_instance_profile, "name", null)
-  #     arn  = lookup(var.iam_instance_profile, "arn", null)
-  #   }
-  # }
-
-  dynamic "instance_market_options" {
-    for_each = var.instance_market_options != null ? [var.instance_market_options] : []
-    content {
-      market_type = instance_market_options.value.market_type
-
-      dynamic "spot_options" {
-        for_each = lookup(instance_market_options.value, "spot_options", null) != null ? [instance_market_options.value.spot_options] : []
-        content {
-          block_duration_minutes         = lookup(spot_options.value, "block_duration_minutes", null)
-          instance_interruption_behavior = lookup(spot_options.value, "instance_interruption_behavior", null)
-          max_price                      = lookup(spot_options.value, "max_price", null)
-          spot_instance_type             = lookup(spot_options.value, "spot_instance_type", null)
-          valid_until                    = lookup(spot_options.value, "valid_until", null)
-        }
-      }
-    }
-  }
-
-  dynamic "license_specification" {
-    for_each = var.license_specifications != null ? [var.license_specifications] : []
-    content {
-      license_configuration_arn = license_specifications.value.license_configuration_arn
-    }
-  }
-
-  dynamic "metadata_options" {
-    for_each = var.metadata_options != null ? [var.metadata_options] : []
-    content {
-      http_endpoint               = lookup(metadata_options.value, "http_endpoint", null)
-      http_tokens                 = lookup(metadata_options.value, "http_tokens", null)
-      http_put_response_hop_limit = lookup(metadata_options.value, "http_put_response_hop_limit", null)
-      http_protocol_ipv6          = lookup(metadata_options.value, "http_protocol_ipv6", null)
-      instance_metadata_tags      = lookup(metadata_options.value, "instance_metadata_tags", null)
-    }
-  }
-
-  dynamic "monitoring" {
-    for_each = var.enable_monitoring != null ? [1] : []
-    content {
-      enabled = var.enable_monitoring
-    }
-  }
-
-  dynamic "network_interfaces" {
-    for_each = var.network_interfaces
-    content {
-      associate_carrier_ip_address = lookup(network_interfaces.value, "associate_carrier_ip_address", null)
-      associate_public_ip_address  = lookup(network_interfaces.value, "associate_public_ip_address", null)
-      delete_on_termination        = lookup(network_interfaces.value, "delete_on_termination", null)
-      description                  = lookup(network_interfaces.value, "description", null)
-      device_index                 = lookup(network_interfaces.value, "device_index", null)
-      interface_type               = lookup(network_interfaces.value, "interface_type", null)
-      ipv4_addresses               = try(network_interfaces.value.ipv4_addresses, [])
-      ipv4_address_count           = lookup(network_interfaces.value, "ipv4_address_count", null)
-      ipv6_addresses               = try(network_interfaces.value.ipv6_addresses, [])
-      ipv6_address_count           = lookup(network_interfaces.value, "ipv6_address_count", null)
-      network_interface_id         = lookup(network_interfaces.value, "network_interface_id", null)
-      private_ip_address           = lookup(network_interfaces.value, "private_ip_address", null)
-      security_groups              = compact(concat(try(network_interfaces.value.security_groups, []), local.security_group_ids))
-      # Set on EKS managed node group, will fail if set here
-      # https://docs.aws.amazon.com/eks/latest/userguide/launch-templates.html#launch-template-basics
-      # subnet_id                  = lookup(network_interfaces.value, "subnet_id", null)
-    }
-  }
-
-  dynamic "placement" {
-    for_each = var.placement != null ? [var.placement] : []
-    content {
-      affinity          = lookup(placement.value, "affinity", null)
-      availability_zone = lookup(placement.value, "availability_zone", null)
-      group_name        = lookup(placement.value, "group_name", null)
-      host_id           = lookup(placement.value, "host_id", null)
-      spread_domain     = lookup(placement.value, "spread_domain", null)
-      tenancy           = lookup(placement.value, "tenancy", null)
-      partition_number  = lookup(placement.value, "partition_number", null)
-    }
-  }
-
-  dynamic "tag_specifications" {
-    for_each = toset(["instance", "volume", "network-interface"])
-    content {
-      resource_type = tag_specifications.key
-      tags          = merge(var.tags, { Name = local.prefix }, var.launch_template_tags)
-    }
-  }
-
-  lifecycle {
-    create_before_destroy = true
-  }
-
-  # Prevent premature access of security group roles and policies by pods that
-  # require permissions on create/destroy that depend on nodes
-
-  tags = var.tags
+module "launch_template" {
+  source                                 = "git::ssh://git@github.com/oozou/terraform-aws-launch-template.git?ref=v1.0.0"
+  count                                  = var.is_create_launch_template ? 1 : 0
+  prefix                                 = var.prefix
+  environment                            = var.environment
+  name                                   = var.name
+  user_data                              = local.platform[var.platform].user_data
+  ebs_optimized                          = var.ebs_optimized
+  ami_id                                 = var.ami_id
+  key_name                               = var.key_name
+  vpc_security_group_ids                 = var.vpc_security_group_ids
+  launch_template_default_version        = var.launch_template_default_version
+  update_launch_template_default_version = var.update_launch_template_default_version
+  disable_api_termination                = var.disable_api_termination
+  kernel_id                              = var.kernel_id
+  ram_disk_id                            = var.ram_disk_id
+  block_device_mappings                  = var.block_device_mappings
+  capacity_reservation_specification     = var.capacity_reservation_specification
+  cpu_options                            = var.cpu_options
+  credit_specification                   = var.credit_specification
+  elastic_gpu_specifications             = var.elastic_gpu_specifications
+  elastic_inference_accelerator          = var.elastic_inference_accelerator
+  enclave_options                        = var.enclave_options
+  instance_market_options                = var.instance_market_options
+  license_specifications                 = var.license_specifications
+  metadata_options                       = var.metadata_options
+  enable_monitoring                      = var.enable_monitoring
+  network_interfaces                     = var.network_interfaces
+  placement                              = var.placement
+  launch_template_tags                   = var.launch_template_tags
+  tags                                   = var.tags
 }